requirements.txt
numpy~=1.25.2
torch~=2.0.1
torchinfo~=1.8.0
setuptools==65.6.3  # https://github.com/pypa/pip/pull/11976
prettytable~=3.8.0
datasets~=2.14.4
PyYAML~=6.0.1
urllib3==1.26.15  # https://github.com/urllib3/urllib3/issues/3020#issuecomment-1557516332
notebook~=6.5.5  # for nbextensions and colab compatibility
jupyter~=1.0.0
jupyter-client~=7.3.1  # colab requirement
jupyter-console~=6.6.3
jupyter-core~=5.3.1
ipython~=7.34.0  # colab requirement
ipython-genutils==0.2.0
ipywidgets==8.1.0
matplotlib~=3.7.2
seaborn~=0.12.2
rich~=13.5.2
jsonlines~=3.1.0
scikit-learn~=1.3.0
dataclasses~=0.6
einops~=0.7.0
tokenizers~=0.14.1
transformers~=4.35.0
huggingface-hub~=0.16.4
rouge-score